feat(admin): updated badge layout for BadgeService
This probably should be evicted to a JSON blob, so updating the badge does not require updating the backend.

A built-in editor is a bit harder due to how shit tc-lib-pdf is with it's bounding box calculation: either TextCell does some wildly imaginary things or the numbers given are ~40% too large.