feat(admin): badge printing (#189)
* fix(dev): use `mc alias` instead of  `mc config`

`mc config` has been deprecated with `mc alias` providing almost the same functionality, albeit different order of params

* feat(admin): add support for generating badges from applications

Also added a custom badge generator, for one-off prints

* fix(admin): show table name in badge list

* feat(admin): add settings for handling font and background uploads

* fix(admin): Drop unused imports

* feat(admin): use storage for resolving uploaded resources

* fix(admin): leding -> leading

* feat(admin): updated badge layout for BadgeService

This probably should be evicted to a JSON blob, so updating the badge does not require updating the backend.

A built-in editor is a bit harder due to how shit tc-lib-pdf is with it's bounding box calculation: either TextCell does some wildly imaginary things or the numbers given are ~40% too large.

* fix(admin): use more sane logic for table numbers from Sharky

* fix(admin): make display name resolving more readable

Also documented order of preference

* fix(admin): handle unused values cleanly

This includes not overriding badge type if it isn't set to custom and skipping table number if has_table is not true.